Consolidate and rename AppointmentPage unit tests
